+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - chapter-9.rb [examples from chapter-9]
 - one-billion-seconds.rb [todo from chapter-9]
+- happy-birthday.rb [todo from chapter-9]
 
 ## [0.0.11] (May 23 2021)